Refactor RoyaltyPayment and GroupRoyaltyDistribution models to use BigIntegerField for block_number. Update DerivativeCreateSerializer to set parent asset as a ForeignKey instead of UUID. Modify IPAssetViewSet to correctly handle license_terms_id for derivative relationships and enhance registration status tracking for derivatives.

+++ apps/assets/models.py
@@ -271,7 +271,7 @@ class RoyaltyPayment(models.Model):
         help_text="Blockchain transaction hash"
     )
 
-    block_number = models.IntegerField(
+    block_number = models.BigIntegerField(
         help_text="Block number where payment occurred"
     )
 
@@ -507,7 +507,7 @@ class GroupRoyaltyDistribution(models.Model):
         help_text="Blockchain transaction hash"
     )
 
-    block_number = models.IntegerField(
+    block_number = models.BigIntegerField(
         help_text="Block number where distribution occurred"
     )
